# api-contract.yaml
openapi: 3.1.0
info:
  title: Crud Market - REST API
  description: |-
    A REST API for a blatantly fake E-Commerce application
  license:
    name: MIT
    url: https://mit-license.org/
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /cart:
    get:
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Filters out the cart items based on the given product name
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: How many cart items to display at once. You can display a maximum of 20 cart items
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: To be used with `limit`. Offsets the query by `limit` times `page`
      description: |-
        Returns all items in the user's cart
      responses:
        "200":
          description: Cart obtained
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "403":
          description: User isn't authenticated
      tags: [cart]
    post:
      description: |-
        Adds the given item to the user's cart. If the item already exists within the cart, increments the existing item's product count instead.
      requestBody:
        description: Contains the information of the product to add
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: number
                  examples:
                    - 0
                    - 1
                    - 2
                    - 3
                    - 4
                quantity:
                  type: number
                  examples:
                    - 1
                    - 2
                    - 3
                    - 4
      responses:
        "201":
          description: Item added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "400":
          description: Attempted to add invalid item
        "403":
          description: User isn't authenticated
      tags: [cart]
    delete:
      description: |-
        Clears all items within the cart without checking out
      responses:
        "204":
          description: Cart cleared successfully
        "403":
          description: User isn't authenticated
      tags: [cart]
  /cart/{itemId}:
    parameters:
      - in: path
        name: itemId
        schema:
          type: integer
        required: true
        description: The ID of a product in the user's cart
    put:
      description: |-
        Updates the quantity of the given product
      requestBody:
        description: Determines the new quantity of items to get
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                  examples:
                    - 1
                    - 2
                    - 3
                    - 4
      responses:
        "200":
          description: Cart item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "204":
          description: Cart item removed when updated. Used when quantity is zero
        "403":
          description: User isn't authenticated
        "404":
          description: Cart item not found
      tags: [cart]
    delete:
      description: |-
        Clears the given item from the user's cart
      responses:
        "204":
          description: Cart item removed successfully
        "403":
          description: User isn't authenticated
        "404":
          description: Cart item not found
      tags: [cart]

  # Checkout Handling:
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        productName:
          type: string
          examples:
            - Weed Eater
            - Soda Carbonator
            - Hair Rake
        quantity:
          type: number
          examples:
            - 1
            - 2
            - 3
            - 4
        price:
          type: number
          examples:
            - 15.57
            - 30.99
            - 45.94
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        price:
          type: number
          examples:
            - 50.47
            - 80.47
            - 102.47
...